requirements.txt
annotated-types==0.6.0
anyio==4.3.0
assemblyai==0.22.0
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
openai==1.12.0
opencv-python==4.9.0.80
pillow==10.2.0
proglog==0.1.10
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
requests==2.31.0
scipy==1.12.0
sniffio==1.3.0
srt==3.5.3
srt_equalizer==0.1.8
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
websockets==12.0
